requirements.txt
Django
pillow
django-embed-video
django-allauth
python-memcached
